Added explicit casting to cleanse warnings

+++ Cerberus/Core/testClass.cpp
@@ -11,12 +11,12 @@ TestClass::TestClass()
 	sprite->SetRenderRect(XMUINT2(128, 128));
 	sprite->SetSpriteSize(XMUINT2(128, 128));
 
-	sprite->SetTint(XMFLOAT4(rand() % 2 * .5, rand() % 2 * .5, rand() % 2 * .5, 0));
+	sprite->SetTint(XMFLOAT4(float(rand() % 2 * .5), float(rand() % 2 * .5), float(rand() % 2 * .5), 0));
 
 	if (rand() % 2)
 		sprite->SetScale(-1, 1, 1);
 
-	timeElapsed = rand() / 100;
+	timeElapsed = float(rand() / 100);
 }
 
 void TestClass::Update(float deltaTime)
@@ -26,7 +26,7 @@ void TestClass::Update(float deltaTime)
 	timeElapsed += deltaTime;
 
 	const uint32_t speed = 24;
-	sprite->SetTextureOffset(XMFLOAT2(round(timeElapsed * speed) * 128, (int(round(timeElapsed * speed) / 5) % 2) * 128));
+	sprite->SetTextureOffset(XMFLOAT2(round(timeElapsed * speed) * 128, float((int(round(timeElapsed * speed) / 5) % 2)) * 128));
 }
 
 TestClass::~TestClass()
